_get_parameters_for_build, changed to divide the main retina parameters from all the other retina parameters.
The main retina parameters are in a list, so that only these main parameters are singled out and used to build the hash and saved to the yaml file. This way, the main parameters are separated from all the other retina_parameters, which are dynamically changed and may not be hashable / serialized in a YAML file.